feat: post PR vote component updates
