+++ app/helpers/error.helper.ts
@@ -1,0 +1,7 @@
+import { BaseError } from "viem";
+
+export default abstract class ErrorHelper {
+  public static processWagmiErrorMessage(error: Error): string {
+    return (error as BaseError).shortMessage || error?.message;
+  }
+}
